+++ R/betwfe_core.R
@@ -1097,6 +1097,7 @@ betwfe_core <- function(
 		T = T,
 		fused = FALSE,
 		calc_ses = q < 1,
+		include_selected = TRUE,
 		alpha = alpha,
 		se_type = se_type,
 		y_final = y_final